refactor additions, cleanup
