+++ packages/core/src/context/types.ts
@@ -1,4 +1,4 @@
-import type { PlatformKey, RenativeConfigFile, RenativeConfigLocal } from '../schema/ts/types';
+import type { PlatformKey, RenativeConfigFile, RenativeConfigLocal } from '../schema/types';
 import type { RnvEngine, RnvEnginePlatform } from '../engines/types';
 import type { OverridesOptions } from '../system/types';
 import type { RnvPlatform } from '../types';
