+++ packages/core/src/index.ts
@@ -32,7 +32,7 @@ export * from './projects/dependencyManager';
 export * from './tasks';
 export * from './tasks/constants';
 
-export * from './schema/zod';
+export * from './schema';
 export * from './context/runtime';
 export * from './templates';
 export * from './integrations';
@@ -54,9 +54,9 @@ export * from './engines/types';
 export * from './context/types';
 export * from './plugins/types';
 export * from './projects/types';
-export * from './schema/ts/types';
+export * from './schema/types';
 export * from './tasks/types';
 export * from './system/types';
 export * from './api/types';
 export * from './types';
-export * from './schema/ts/types';
+export * from './schema/types';
