+++ packages/core/src/schema/configFiles/appConfig.ts
