+++ packages/core/src/schema/configFiles/global.ts
