+++ packages/core/src/schema/configFiles/local.ts
