+++ packages/core/src/schema/configFiles/plugin.ts
