+++ packages/core/src/schema/configFiles/pluginTemplates.ts
