+++ packages/core/src/schema/configFiles/project.ts
