+++ packages/core/src/schema/platforms/electron.ts
