+++ packages/core/src/schema/platforms/tizen.ts
