+++ packages/core/src/schema/platforms/web.ts
@@ -1,0 +1,10 @@
+import { z } from 'zod';
+
+export const PlatformWeb = z.object({
+    // environment: {
+    //     type: 'string',
+    // },
+    // devServerHost: {
+    //     type: 'string',
+    // },
+});
