+++ packages/core/src/schema/platforms/webpack.ts
