+++ packages/core/src/schema/platforms/windows.ts
