+++ packages/core/src/schema/plugins/android.ts
@@ -1,13 +1,14 @@
 import { z } from 'zod';
-import { PluginShared } from './base';
-import { PlatformSharedAndroid } from '../platforms/androidBase';
+import { PluginPlatformBase } from './platformBase';
+import { TemplateAndroidShared } from '../platforms/templateAndroid';
 
-export const PluginAndroid = PluginShared.merge(PlatformSharedAndroid).merge(
+export const PluginAndroid = PluginPlatformBase.merge(
     z.object({
         projectName: z.optional(z.string()),
         skipLinking: z.optional(z.boolean()),
         skipImplementation: z.optional(z.boolean()),
         implementation: z.optional(z.string()),
         package: z.optional(z.string()),
+        templateAndroid: z.optional(TemplateAndroidShared),
     })
 );
